UIAPPS-167 Separate 1st and 3rd party data
Motivation
----------

- Jira Issue: https://datadoghq.atlassian.net/browse/UIAPPS-167
- The kitchen sink is going to be mostly the same as the previous Apps,
    so we're taking a second to clean up some of this stuff.

Changes
-------

- We separate the 1st-party and 3rd-party data from the App stuff. This
    should help with understanding what all the pieces do from a high
    level.

Testing
-------

- This is mostly shuffling files. The behavior should be the same.

Releases
--------

Choose one:

-   [x] No release is necessary.
        If you're only updating examples/documentation, this is likely what you want.
-   [ ] All packages that need a release have a changeset.
