+++ examples/custom-widget-with-1st-party-data/src/1st-party/monitor/index.ts
@@ -1,0 +1,1 @@
+export * from './monitor';
