+++ examples/custom-widget-with-3rd-party-data/src/3rd-party/login/login.tsx
