+++ examples/custom-widget-with-3rd-party-data/src/3rd-party/post/post.ts
