+++ examples/widget-context-menu-with-3rd-party-data/src/3rd-party/login/login.tsx
