+++ examples/widget-context-menu-with-3rd-party-data/src/3rd-party/user/index.ts
